+++ CHANGELOG.md
@@ -6,9 +6,22 @@
 
 ### New Features
 
+- Added support for subscripts
+
+## 0.10.1
+
+* When installing Sourcery via CocoaPods, the unneeded `file.zip` is not kept in `Pods/Sourcery/` anymore _(freeing ~12MB on each install of Sourcery made via CocoaPods!)_.  
+
+## 0.10.0
+
+### New Features
+
+- Added test for count Stencil filter
+- Added new reversed Stencil filter
+- Added new isEmpty Stencil filter
+- Added new sorted and sortedDescending Stencil filters. This can sort arrays by calling e.g. `protocol.allVariables|sorted:"name"`
 - Added new toArray Stencil filter
 - Added a console warning when a yaml is available but any parameter between 'sources', templates', 'forceParse', 'output' are provided
-- Added support for subscripts
 
 ### Internal changes